Search issue fix
The search plugin was creating build issues because one of its dependencies (Cheerio) was upgraded and the search plugin package did not contain the necessary changes pertaining to this upgrade.

On the search plugin repo, until the PR containing changes related to Cheerio version is merged, we are adding a workaround in package.json as specified in https://github.com/cmfcmf/docusaurus-search-local/pull/218

After the search plugin package is updated, the 'overrides' may not be necessary in package.json.

Also, the search plugin version is reverted to 1.2 as it is compatible with Docusaurus v2.